+++ crates/gem_bitcoin/src/typeshare/transaction.rs
@@ -1,11 +1,13 @@
 use serde::{Deserialize, Serialize};
 use typeshare::typeshare;
 
+use super::UInt64;
+
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[typeshare(swift = "Sendable")]
 pub struct BitcoinTransaction {
     #[serde(rename = "blockHeight")]
-    pub block_height: i32,
+    pub block_height: UInt64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
